split up spec tests into mainnet and minimal, skip any minimal test that are failing without --define ssz=minimal
