+++ beacon-chain/core/blocks/spectest/attestation_mainnet_test.go
@@ -1,0 +1,9 @@
+package spectest
+
+import (
+	"testing"
+)
+
+func TestAttestationMainnet(t *testing.T) {
+	runAttestationTest(t, "attestation_mainnet.yaml")
+}
