+++ beacon-chain/core/blocks/spectest/attestation_minimal_test.go
@@ -1,0 +1,9 @@
+package spectest
+
+import (
+	"testing"
+)
+
+func TestAttestationMinimal(t *testing.T) {
+	runAttestationTest(t, "attestation_minimal.yaml")
+}
