+++ beacon-chain/core/blocks/spectest/block_header_minimal_test.go
@@ -1,0 +1,15 @@
+package spectest
+
+import (
+	"testing"
+
+	"github.com/bazelbuild/rules_go/go/tools/bazel"
+)
+
+func TestBlockHeaderMinimal(t *testing.T) {
+	filepath, err := bazel.Runfile(blkHeaderPrefix + "block_header_minimal.yaml")
+	if err != nil {
+		t.Fatal(err)
+	}
+	runBlockHeaderTest(t, filepath)
+}
