+++ beacon-chain/core/blocks/spectest/block_processing_mainnet_test.go
@@ -1,0 +1,9 @@
+package spectest
+
+import (
+	"testing"
+)
+
+func TestBlockProcessingMainnetYaml(t *testing.T) {
+	runBlockProcessingTest(t, "sanity_blocks_mainnet.yaml")
+}
