+++ beacon-chain/core/blocks/spectest/deposit_test.go
@@ -5,13 +5,14 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/bazelbuild/rules_go/go/tools/bazel"
 	"github.com/prysmaticlabs/prysm/beacon-chain/core/blocks"
 	"github.com/prysmaticlabs/prysm/beacon-chain/core/helpers"
 	"github.com/prysmaticlabs/prysm/beacon-chain/core/state/stateutils"
 	"github.com/prysmaticlabs/prysm/shared/params/spectest"
 	"github.com/prysmaticlabs/prysm/shared/testutil"
 )
+
+const depositPrefix = "tests/operations/deposit/"
 
 func runDepositTest(t *testing.T, filename string) {
 	file, err := ioutil.ReadFile(filename)
@@ -59,22 +60,4 @@ func runDepositTest(t *testing.T, filename string) {
 			}
 		})
 	}
-}
-
-var depositPrefix = "tests/operations/deposit/"
-
-func TestDepositMinimalYaml(t *testing.T) {
-	filepath, err := bazel.Runfile(depositPrefix + "deposit_minimal.yaml")
-	if err != nil {
-		t.Fatal(err)
-	}
-	runDepositTest(t, filepath)
-}
-
-func TestDepositMainnetYaml(t *testing.T) {
-	filepath, err := bazel.Runfile(depositPrefix + "deposit_mainnet.yaml")
-	if err != nil {
-		t.Fatal(err)
-	}
-	runDepositTest(t, filepath)
 }
