+++ beacon-chain/core/blocks/spectest/proposer_slashing_mainnet_test.go
@@ -1,0 +1,15 @@
+package spectest
+
+import (
+	"testing"
+
+	"github.com/bazelbuild/rules_go/go/tools/bazel"
+)
+
+func TestProposerSlashingMainnet(t *testing.T) {
+	filepath, err := bazel.Runfile(proposerSlashingPrefix + "proposer_slashing_mainnet.yaml")
+	if err != nil {
+		t.Fatal(err)
+	}
+	runProposerSlashingTest(t, filepath)
+}
