+++ beacon-chain/core/blocks/spectest/proposer_slashing_test.go
@@ -4,7 +4,6 @@ import (
 	"io/ioutil"
 	"testing"
 
-	"github.com/bazelbuild/rules_go/go/tools/bazel"
 	"github.com/gogo/protobuf/proto"
 	"github.com/prysmaticlabs/prysm/beacon-chain/core/blocks"
 	"github.com/prysmaticlabs/prysm/beacon-chain/core/helpers"
@@ -13,6 +12,8 @@ import (
 	"github.com/prysmaticlabs/prysm/shared/testutil"
 	"gopkg.in/d4l3k/messagediff.v1"
 )
+
+const proposerSlashingPrefix = "tests/operations/proposer_slashing/"
 
 func runProposerSlashingTest(t *testing.T, filename string) {
 	file, err := ioutil.ReadFile(filename)
@@ -59,22 +60,4 @@ func runProposerSlashingTest(t *testing.T, filename string) {
 			}
 		})
 	}
-}
-
-var proposerSlashingPrefix = "tests/operations/proposer_slashing/"
-
-func TestProposerSlashingMinimal(t *testing.T) {
-	filepath, err := bazel.Runfile(proposerSlashingPrefix + "proposer_slashing_minimal.yaml")
-	if err != nil {
-		t.Fatal(err)
-	}
-	runProposerSlashingTest(t, filepath)
-}
-
-func TestProposerSlashingMainnet(t *testing.T) {
-	filepath, err := bazel.Runfile(proposerSlashingPrefix + "proposer_slashing_mainnet.yaml")
-	if err != nil {
-		t.Fatal(err)
-	}
-	runProposerSlashingTest(t, filepath)
 }
