+++ beacon-chain/core/blocks/spectest/transfer_test.go
@@ -4,7 +4,6 @@ import (
 	"io/ioutil"
 	"testing"
 
-	"github.com/bazelbuild/rules_go/go/tools/bazel"
 	"github.com/gogo/protobuf/proto"
 	"github.com/prysmaticlabs/prysm/beacon-chain/core/blocks"
 	"github.com/prysmaticlabs/prysm/beacon-chain/core/helpers"
@@ -13,6 +12,8 @@ import (
 	"github.com/prysmaticlabs/prysm/shared/testutil"
 	"gopkg.in/d4l3k/messagediff.v1"
 )
+
+const transferPrefix = "tests/operations/transfer/"
 
 func runTransferTest(t *testing.T, filename string) {
 	file, err := ioutil.ReadFile(filename)
@@ -59,24 +60,4 @@ func runTransferTest(t *testing.T, filename string) {
 			}
 		})
 	}
-}
-
-var transferPrefix = "tests/operations/transfer/"
-
-func TestTransferMinimal(t *testing.T) {
-	t.Skip("Transfer tests are disabled. See https://github.com/ethereum/eth2.0-specs/pull/1238#issuecomment-507054595")
-	filepath, err := bazel.Runfile(transferPrefix + "transfer_minimal.yaml")
-	if err != nil {
-		t.Fatal(err)
-	}
-	runTransferTest(t, filepath)
-}
-
-func TestTransferMainnet(t *testing.T) {
-	t.Skip("Transfer tests are disabled. See https://github.com/ethereum/eth2.0-specs/pull/1238#issuecomment-507054595")
-	filepath, err := bazel.Runfile(transferPrefix + "transfer_mainnet.yaml")
-	if err != nil {
-		t.Fatal(err)
-	}
-	runTransferTest(t, filepath)
 }
