+++ beacon-chain/core/blocks/spectest/voluntary_exit_minimal_test.go
@@ -1,0 +1,15 @@
+package spectest
+
+import (
+	"testing"
+
+	"github.com/bazelbuild/rules_go/go/tools/bazel"
+)
+
+func TestVoluntaryExitMinimal(t *testing.T) {
+	filepath, err := bazel.Runfile(exitPrefix + "voluntary_exit_mainnet.yaml")
+	if err != nil {
+		t.Fatal(err)
+	}
+	runVoluntaryExitTest(t, filepath)
+}
