+++ beacon-chain/core/epoch/spectest/BUILD.bazel
@@ -8,16 +8,49 @@ go_library(
     deps = ["//proto/beacon/p2p/v1:go_default_library"],
 )
 
-go_test(
+test_suite(
     name = "go_default_test",
-    size = "medium",
-    srcs = [
-        "crosslink_test.go",
-        "final_updates_test.go",
-        "justification_and_finalization_test.go",
-        "registry_test.go",
-        "slashings_test.go",
+    tags = ["spectest"],
+    tests = [
+        ":go_mainnet_test",
+        ":go_minimal_test",
     ],
+)
+
+go_test(
+    name = "go_mainnet_test",
+    size = "small",
+    srcs = glob(
+        ["*_test.go"],
+        exclude = ["*_minimal_test.go"],
+    ),
+    data = [
+        "@eth2_spec_tests//:test_data",
+    ],
+    embed = [":go_default_library"],
+    shard_count = 4,
+    tags = [
+        "spectest",
+    ],
+    deps = [
+        "//beacon-chain/core/epoch:go_default_library",
+        "//beacon-chain/core/helpers:go_default_library",
+        "//proto/beacon/p2p/v1:go_default_library",
+        "//shared/params/spectest:go_default_library",
+        "//shared/testutil:go_default_library",
+        "@com_github_gogo_protobuf//proto:go_default_library",
+        "@in_gopkg_d4l3k_messagediff_v1//:go_default_library",
+        "@io_bazel_rules_go//go/tools/bazel:go_default_library",
+    ],
+)
+
+go_test(
+    name = "go_minimal_test",
+    size = "small",
+    srcs = glob(
+        ["*_test.go"],
+        exclude = ["*_mainnet_test.go"],
+    ),
     data = [
         "@eth2_spec_tests//:test_data",
     ],
