+++ beacon-chain/core/epoch/spectest/final_updates_minimal_test.go
@@ -1,0 +1,15 @@
+package spectest
+
+import (
+	"testing"
+
+	"github.com/bazelbuild/rules_go/go/tools/bazel"
+)
+
+func TestFinalUpdatesMinimal(t *testing.T) {
+	filepath, err := bazel.Runfile(finalUpdatesPrefix + "final_updates_minimal.yaml")
+	if err != nil {
+		t.Fatal(err)
+	}
+	runFinalUpdatesTests(t, filepath)
+}
