+++ beacon-chain/core/epoch/spectest/justification_and_finalization_test.go
@@ -6,7 +6,6 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/bazelbuild/rules_go/go/tools/bazel"
 	"github.com/gogo/protobuf/proto"
 	"github.com/prysmaticlabs/prysm/beacon-chain/core/epoch"
 	"github.com/prysmaticlabs/prysm/beacon-chain/core/helpers"
@@ -15,6 +14,8 @@ import (
 	"github.com/prysmaticlabs/prysm/shared/testutil"
 	"gopkg.in/d4l3k/messagediff.v1"
 )
+
+const justificationAndFinalizationPrefix = "tests/epoch_processing/justification_and_finalization/"
 
 // This is a subset of state.ProcessEpoch. The spec test defines input data for
 // `justification_and_finalization` only.
@@ -98,24 +99,4 @@ func runJustificationAndFinalizationTests(t *testing.T, filename string) {
 			}
 		})
 	}
-}
-
-const justificationAndFinalizationPrefix = "tests/epoch_processing/justification_and_finalization/"
-
-func TestJustificationAndFinalizationMinimal(t *testing.T) {
-	// TODO(#2891): Verify with ETH2 spec test.
-	t.Skip("The input data fails preconditions for matching attestations in the state for the current epoch.")
-	filepath, err := bazel.Runfile(justificationAndFinalizationPrefix + "justification_and_finalization_minimal.yaml")
-	if err != nil {
-		t.Fatal(err)
-	}
-	runJustificationAndFinalizationTests(t, filepath)
-}
-
-func TestJustificationAndFinalizationMainnet(t *testing.T) {
-	filepath, err := bazel.Runfile(justificationAndFinalizationPrefix + "justification_and_finalization_mainnet.yaml")
-	if err != nil {
-		t.Fatal(err)
-	}
-	runJustificationAndFinalizationTests(t, filepath)
 }
