+++ beacon-chain/core/epoch/spectest/registry_test.go
@@ -5,11 +5,12 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/bazelbuild/rules_go/go/tools/bazel"
 	"github.com/prysmaticlabs/prysm/beacon-chain/core/epoch"
 	"github.com/prysmaticlabs/prysm/shared/params/spectest"
 	"github.com/prysmaticlabs/prysm/shared/testutil"
 )
+
+const registryUpdatesPrefix = "tests/epoch_processing/registry_updates/"
 
 func runRegisteryProcessingTests(t *testing.T, filename string) {
 	file, err := ioutil.ReadFile(filename)
@@ -42,22 +43,4 @@ func runRegisteryProcessingTests(t *testing.T, filename string) {
 			}
 		})
 	}
-}
-
-const registryUpdatesPrefix = "tests/epoch_processing/registry_updates/"
-
-func TestRegistryProcessingMinimal(t *testing.T) {
-	filepath, err := bazel.Runfile(registryUpdatesPrefix + "registry_updates_minimal.yaml")
-	if err != nil {
-		t.Fatal(err)
-	}
-	runRegisteryProcessingTests(t, filepath)
-}
-
-func TestRegistryProcessingMainnet(t *testing.T) {
-	filepath, err := bazel.Runfile(registryUpdatesPrefix + "registry_updates_mainnet.yaml")
-	if err != nil {
-		t.Fatal(err)
-	}
-	runRegisteryProcessingTests(t, filepath)
 }
