+++ beacon-chain/core/epoch/spectest/slashings_minimal_test.go
@@ -1,0 +1,15 @@
+package spectest
+
+import (
+	"testing"
+
+	"github.com/bazelbuild/rules_go/go/tools/bazel"
+)
+
+func TestSlashingsMinimal(t *testing.T) {
+	filepath, err := bazel.Runfile(slashingsPrefix + "slashings_minimal.yaml")
+	if err != nil {
+		t.Fatal(err)
+	}
+	runSlashingsTests(t, filepath)
+}
