+++ beacon-chain/core/state/spectest/BUILD.bazel
@@ -15,13 +15,49 @@ go_library(
     ],
 )
 
-go_test(
+test_suite(
     name = "go_default_test",
-    size = "large",
-    srcs = [
-        "genesis_test.go",
-        "slot_processing_test.go",
+    tags = ["spectest"],
+    tests = [
+        ":go_mainnet_test",
+        ":go_minimal_test",
     ],
+)
+
+go_test(
+    name = "go_mainnet_test",
+    size = "medium",
+    srcs = glob(
+        ["*_test.go"],
+        exclude = ["*_minimal_test.go"],
+    ),
+    data = [
+        "@eth2_spec_tests//:test_data",
+    ],
+    embed = [":go_default_library"],
+    shard_count = 2,
+    tags = ["spectest"],
+    deps = [
+        "//beacon-chain/core/helpers:go_default_library",
+        "//beacon-chain/core/state:go_default_library",
+        "//proto/eth/v1alpha1:go_default_library",
+        "//shared/params:go_default_library",
+        "//shared/params/spectest:go_default_library",
+        "//shared/testutil:go_default_library",
+        "@com_github_gogo_protobuf//proto:go_default_library",
+        "@com_github_prysmaticlabs_go_ssz//:go_default_library",
+        "@in_gopkg_d4l3k_messagediff_v1//:go_default_library",
+        "@io_bazel_rules_go//go/tools/bazel:go_default_library",
+    ],
+)
+
+go_test(
+    name = "go_minimal_test",
+    size = "small",
+    srcs = glob(
+        ["*_test.go"],
+        exclude = ["*_mainnet_test.go"],
+    ),
     data = [
         "@eth2_spec_tests//:test_data",
     ],
