+++ beacon-chain/core/state/spectest/slot_processing_mainnet_test.go
@@ -1,0 +1,9 @@
+package spectest
+
+import (
+	"testing"
+)
+
+func TestSlotProcessingMainnet(t *testing.T) {
+	runSlotProcessingTests(t, "sanity_slots_mainnet.yaml")
+}
