+++ proto/testing/ssz_mainnet_compatibility_test.go
@@ -15,7 +15,7 @@ import (
 	"github.com/prysmaticlabs/prysm/shared/testutil"
 )
 
-func TestYamlStatic(t *testing.T) {
+func TestYamlStatic_Mainnet(t *testing.T) {
 	topPath := "tests/ssz_static/core/"
 	yamlFileNames := []string{
 		"ssz_mainnet_random.yaml",
@@ -35,11 +35,11 @@ func TestYamlStatic(t *testing.T) {
 		if err := yaml.Unmarshal(file, s); err != nil {
 			t.Fatalf("Failed to unmarshal: %v", err)
 		}
-		runTestCases(t, s)
+		runTestCases_Mainnet(t, s)
 	}
 }
 
-func runTestCases(t *testing.T, s *sszspectest.SszMainnetTest) {
+func runTestCases_Mainnet(t *testing.T, s *sszspectest.SszMainnetTest) {
 	for _, testCase := range s.TestCases {
 		if !testutil.IsEmpty(testCase.Attestation.Value) {
 			p := &ethpb.Attestation{}
